+++ src/index.js
@@ -141,3 +141,8 @@ export { Collapse } from "./components/Collapse/Collapse";
 export { Download } from "./components/Download/Download";
 
 export { TopNav } from "./components/TopNav/TopNav";
+
+/*
+ * Date Modified component
+ */
+export { Date } from "./components/Date/Date";