+++ CHANGELOG.md
@@ -45,6 +45,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 - TopNav Component
 - Header Update
 - Layout Container Update
+- Date Modified Component
 
 ### Changed